// src/types/Api.ts
export interface SaveListMaterial {
  Material: string;
  Location: string;
  QTY: number;
}
export interface LitsTableResponse {
  Material: string;
  ProductCode: string;
  Location: string;
  QTY: number;
}
export interface GetMaterial {
  Material: string;
  ProductCode: string;
  Description: string;
}